+++ CHANGELOG.md
@@ -342,3 +342,11 @@ regla de "En particular, en caso de disponer de una sola ficha que puede moverse
  dados_disponibles pasado como atributo de backgammon
 ### Arreglado
 ###
+
+## [2.4.1] - 27-10-2025 ##
+### Agregado
+cartel para mostrar dados sacados primer turno en ui
+### Cambiado
+Se agrega interfaz ICartelUI para implementar dip en Cartel_UI
+### Arreglado
+###